refactor: Entity 클래스 이름 수정
+ TimeBase column명 snake_case 규칙 적용
