+++ src/main/java/com/FitToMe/project/Entity/User.java
@@ -1,6 +1,7 @@
-package com.FitToMe.project.Entity.User;
+package com.FitToMe.project.Entity;
 
 import lombok.*;
+
 import javax.persistence.*;
 
 @Entity
