+++ src/main/java/com/FitToMe/project/Service/Post/PostDeleteService.java
@@ -1,6 +1,6 @@
 package com.FitToMe.project.Service.Post;
 
-import com.FitToMe.project.Entity.Post.PostEntity;
+import com.FitToMe.project.Entity.Post;
 import com.FitToMe.project.Repository.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -14,7 +14,7 @@ public class PostDeleteService {
     private final PostRepository postRepository;
 
     public boolean deletePost(Long postId) throws RuntimeException {
-        PostEntity post = postRepository.findById(postId)
+        Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new RuntimeException("존재하지 않는 게시글입니다"));
 
         postRepository.deleteById(postId);
