+++ src/main/java/com/FitToMe/project/Service/Post/PostModifyService.java
@@ -1,6 +1,6 @@
 package com.FitToMe.project.Service.Post;
 
-import com.FitToMe.project.Entity.Post.PostEntity;
+import com.FitToMe.project.Entity.Post;
 import com.FitToMe.project.Repository.PostRepository;
 import com.FitToMe.project.Request.PostModifyRequest;
 import lombok.RequiredArgsConstructor;
@@ -15,7 +15,7 @@ public class PostModifyService {
     private final PostRepository postRepository;
 
     public com.FitToMe.project.DTO.PostDTO updatePost(Long postId, PostModifyRequest postModifyRequest) throws IllegalArgumentException {    //예외처리 check
-        PostEntity post = postRepository.findById(postId)
+        Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 게사판은 존재하지 않습니다"));
 
         if (postModifyRequest.getTitle() != null) {
