+++ src/main/java/com/FitToMe/project/Service/Post/PostRegisterService.java
@@ -1,7 +1,7 @@
 package com.FitToMe.project.Service.Post;
 
 import com.FitToMe.project.DTO.PostDTO;
-import com.FitToMe.project.Entity.Post.PostEntity;
+import com.FitToMe.project.Entity.Post;
 import com.FitToMe.project.Repository.PostRepository;
 import com.FitToMe.project.Request.PostRegisterRequest;
 import lombok.RequiredArgsConstructor;
@@ -16,6 +16,6 @@ public class PostRegisterService {
     private final PostRepository postRepository;
 
     public PostDTO createPost(PostRegisterRequest postRegisterRequest) {
-        return new PostDTO(postRepository.save(new PostEntity(postRegisterRequest)));
+        return new PostDTO(postRepository.save(new Post(postRegisterRequest)));
     }
 }
