+++ src/main/java/com/FitToMe/project/Service/Post/PostStatusService.java
@@ -1,7 +1,7 @@
 package com.FitToMe.project.Service.Post;
 
 import com.FitToMe.project.DTO.PostDTO;
-import com.FitToMe.project.Entity.Post.PostEntity;
+import com.FitToMe.project.Entity.Post;
 import com.FitToMe.project.Repository.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,19 +18,19 @@ public class PostStatusService {
     private final PostRepository postRepository;
 
     public PostDTO findOne(Long postId) throws NullPointerException {
-        PostEntity post = postRepository.findById(postId)
+        Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new NullPointerException("해당 id 게시글이 존재하지 않습니다"));
         return new PostDTO(post);
     }
 
     public PostDTO findOne(String postTitle) throws NullPointerException {
-        PostEntity post = postRepository.findByTitle(postTitle)
+        Post post = postRepository.findByTitle(postTitle)
                 .orElseThrow(() -> new NullPointerException("해당 title 게시글이 존재하지 않습니다"));
         return new PostDTO(post);
     }
 
     public List<PostDTO> findAll() {
-        List<PostEntity> posts = postRepository.findAll();
+        List<Post> posts = postRepository.findAll();
         return posts.stream().map(PostDTO::new).collect(Collectors.toList());
     }
 }
